anov	R Documentation

Tabla anova de la regresión

Description

Realiza una tabla Anova, nos muestra el coeficiente de determinación y el contraste de regresión

Usage


anov(x,y,alpha)

Arguments

x
un objeto matriz de las variables independientes

y
un objeto vector de la variable dependiente

alpha
un objeto numérico que indique el nivel de significación del contraste

Details

Calcula la tabla Anova sobre las variables con las que estamos trabajando, además unas vez que obtiene la tabla, calcula el coeficiente de determinación.

Y por último compara el p-valor con el alpha para ver si se rechaza o no el contraste.

Value

La función anov nos muestra la tabla que obtendríamos con las variables que introducimos, luego nos muestra el coeficiente de determinación.

Y por último nos muestra si se rechaza o acepta el contraste de regresión en función del p-valor.

Author(s)

Julio García Bustos, Juan Luis Martínez Gualda, Juan José De la Hoz Gómez y Eduardo Arroyo Galán

References

Formulario Regresión Lineal

Ejemplos Regresión Lineal

Examples


x<-matrix(c(3,1,5,2,4,5,4,6,4,6),ncol=2)
y<-c(3,1,8,3,5)
alpha<-0.05
anov(x,y,alpha)