contraste	R Documentation

Contrastes de hipótesis sobre los parámetros del modelo

Description

Resuelve los contrastes de nulidad de los parámetros del modelo y contrastes sobre un subvector del vector de parámetros.

Usage


contraste(x,y,alpha,R,r)

Arguments

x
un objeto matriz de las variables independientes

y
un objeto vector de la variable dependiente

alpha
un objeto numérico que indique el nivel de significación del contraste

R
un objeto matriz que contiene restricciones lineales

r
un objeto vector que contiene restricciones lineales

Details

Resuelve contrastes definidos por las restricciones lineales que le indiquemos sobre los parámetros de nuestro modelo de regresión lineal múltiple.

Además de reestimar el modelo bajo dichas restricciones lineales.

Value

La función contraste nos muestra la predicción de la variable junto con la predicción de su media.

Además nos muestra un invervalo de confianza de dichas predicciones.

Author(s)

Julio García Bustos, Juan Luis Martínez Gualda, Juan José De la Hoz Gómez y Eduardo Arroyo Galán

References

Formulario Regresión Lineal

Ejemplos Regresión Lineal

Examples


x<-matrix(c(3,1,5,2,4,5,4,6,4,6),ncol=2)
y<-c(3,1,8,3,5)
alpha<-0.05
R<-matrix(c(1,0,0,1,0,-2),ncol=3)
r<-c(3,0)

contraste(x,y,alpha,R,r)