estimacion	R Documentation

Estimación de los parámetros del modelo

Description

Estima los parámetros del modelo, además de la varianza de los errores

Usage


estimacion(x,y)

Arguments

x
Un objeto matriz de variables independientes

y
un objeto vector de la variable dependiente

Details

Una vez introduccido las variables dependientes mediante una matriz y la variable independiente mendiante un vector,realiza una estimación de los parámetros beta mediante el uso de las fórmulas, además de calcular la varianza de errores.

Value

La función estimacion nos muestra mediante una lista los coeficientes del modelo, es decir, los betas obtenidos y el valor de la varianza de errores.

Author(s)

Julio García Bustos, Juan Luis Martínez Gualda, Juan José De la Hoz Gómez y Eduardo Arroyo Galán

References

Formulario Regresión Lineal

Ejemplos Regresión Lineal

Examples


x<-matrix(c(3,1,5,2,4,5,4,6,4,6),ncol=2)
y<-c(3,1,8,3,5)
estimacion(x,y)