intervalo	R Documentation

Intervalos de confianza para los parámetros del modelo

Description

Proporciona intervalos de confianza para los parámetros del modelo de regresión lineal múltiple a un nivel de confianza.

Usage


intervalo(x, y, alpha)

Arguments

x
un objeto matriz de las variables independientes

y
un objeto vector de la variable dependiente

alpha
un objeto numérico que indique el nivel de significación del contraste

Details

Realiza los intervalos de confianza devolviendo los extremos inferiores y superiores de los coeficientes del modelo y de la varianza de los errores.

Value

La función intervalo nos muestra una lista cuyo primer elemento es el intervalo de los coeficientes Beta (en forma matricial) y cuyo segundo elemento es el de la varianza de los errores.

Author(s)

Julio García Bustos, Juan Luis Martínez Gualda, Juan José De la Hoz Gómez y Eduardo Arroyo Galán

References

Formulario Regresión Lineal

Ejemplos Regresión Lineal

Examples


x<-matrix(c(3,1,5,2,4,5,4,6,4,6),ncol=2)
y<-c(3,1,8,3,5)
alpha<-0.05

intervalo(x,y,alpha)