Nulidad	R Documentation

Contraste de nulidad y sobre un subvector

Description

Resuelve los contrastes de nulidad de los parámetros del modelo y contrastes sobre un subvector del vector de parámetros.

Usage


nulidad(x,y,i,alpha)

Arguments

x
un objeto matriz de las variables independientes

y
un objeto vector de la variable dependiente

i
un objeto numérico que indique el coeficiente de beta sobre el que se quiere hacer el contraste

j
un objeto numérico que indique el coeficiente de beta sobre el que se quiere hacer el contraste

alpha
un objeto númerico que indique el nivel de significación del contraste

beta0
un objeto vector que indique el valor de los beta en el contraste del subvector que queremos realizar

Details

Realiza el contraste de nulidad sobre los betas que indiquemos de un modelo de regresión lineal múltiple. Para ello debemos de indicar primero las variables independientes mediantes una matriz, luego el vector que contenga a la variable dependiente.

Además debemos de indicar el número del beta sobre el que queremos realizar el contraste. También hay que indicar el valor del alpha de nuestro contraste.

También realiza un contraste de un subvector de los betas que nosotros le indiquemos.

En pantalla nos mostrará si se acepta o se rechaza ambas hipótesis nulas, diferenciando entre el contraste de nulidad y del subvector.

Value

La función nulidad nos muestra “Se rechaza la hipótesis nula” si el estadístico Texp de contraste en valor absoluto es estrictamente mayor que el valor de la t-student asociado en el contraste de nulidad primero.

Y nos muestra “No se rechaza la hipótesis nula” si el estadístico de contraste en valor absoluto es menor o igual que el valor de la “t-student” asociado.

Además nos muestra “Se rechaza la hipótesis nula” si el estadístico de contraste Fexp es estrictamente mayor que el valor de la F-snedecor.

Y si el estadístico de contraste Fexp es menor o igual que el valor de la F-snedecor asociado nos muestra “No se rechaza la hipótesis nula”.

Author(s)

Julio García Bustos, Juan Luis Martínez Gualda, Juan José De la Hoz Gómez y Eduardo Arroyo Galán

References

Formulario Regresión Lineal

Ejemplos Regresión Lineal

Examples


x<-matrix(c(3,1,5,2,4,5,4,6,4,6),ncol=2)
y<-c(3,1,8,3,5)
alpha<-0.05
i<-1
j<-2
beta0<-c(1,-2)
nulidad(x,y,i,j,alpha,beta0)