Prediccion	R Documentation

Predecir la variable dependiente y su media

Description

Predice la variable dependediente,su media y determina el intervalo de confianza de dichas predicciones

Usage


prediccion(x,y,x0,alpha)

Arguments

x
un objeto matriz de las variables independientes

y
un objeto vector de la variable dependiente

x0
un vector que contiene las nuevas observaciones de las variables independientes

alpha
un objeto numérico que indique el nivel de significación del contraste

Details

Predice la variable con los nuevos valores observado de las variables independiente predice la variable dependiente y su media.

Además realiza los intervalos de confianza para las predicciones obtenidas al nivel de confianza que le indiquemos.

Value

La función prediccion nos muestra la preddicción de la variable junto con la preddicción de su media.

Ademas nos muestra un invervalo de confianza de dichas preddicciones.

Author(s)

Julio García Bustos, Juan Luis Martínez Gualda, Juan José De la Hoz Gómez y Eduardo Arroyo Galán

References

Formulario Regresión Lineal

Ejemplos Regresión Lineal

Examples


x<-matrix(c(3,1,5,2,4,5,4,6,4,6),ncol=2)
y<-c(3,1,8,3,5)
x0<-c(1,4.75,5.25)
alpha<-0.05

prediccion(x,y,x0,alpha)